9518 - WIP cleaned up tests for setNoticesForCalendaredTrialSessionInteractor
